+++ TERMS/cte_termfunc.h
@@ -66,9 +66,9 @@ bool   TermStructEqual(Term_p t1, Term_p t2);
 bool   TermStructEqualNoDeref(Term_p t1, Term_p t2);
 bool   TermStructEqualDeref(Term_p t1, Term_p t2, DerefType deref_1, DerefType deref_2);
 
-int    TermStructWeightCompare(Term_p t1, Term_p t2);
+long   TermStructWeightCompare(Term_p t1, Term_p t2);
 
-int    TermLexCompare(Term_p t1, Term_p t2);
+long   TermLexCompare(Term_p t1, Term_p t2);
 
 bool   TermIsSubterm(Term_p super, Term_p test, DerefType deref);
